+++ Gobchat.App/Core/Config/IGobchatConfigManager.cs
@@ -1,5 +1,5 @@
 ﻿/*******************************************************************************
- * Copyright (C) 2019-2020 MarbleBag
+ * Copyright (C) 2019-2021 MarbleBag
  *
  * This program is free software: you can redistribute it and/or modify it under
  * the terms of the GNU Affero General Public License as published by the Free
@@ -68,6 +68,8 @@ namespace Gobchat.Core.Config
         bool HasProperty(string key);
 
         void SetProperty(string key, object value);
+
+        void DeleteProperty(string key);
 
         void SetGlobalProperty(string key, object value);