+++ Gobchat.App/resources/ui/gobchat/LocaleManager.js
@@ -1,5 +1,5 @@
 ﻿/*******************************************************************************
- * Copyright (C) 2019-2020 MarbleBag
+ * Copyright (C) 2019-2021 MarbleBag
  *
  * This program is free software: you can redistribute it and/or modify it under
  * the terms of the GNU Affero General Public License as published by the Free
@@ -57,6 +57,36 @@ var Gobchat = (function (Gobchat) {
     class LocaleManager {
         setLocale(locale) {
             this._locale = locale
+        }
+
+        async load(keys, params, lookup, language) {
+            keys = [].concat(keys)
+            language = language || this._locale
+
+            var missingKeys = keys
+            if (lookup)
+                missingKeys = keys.filter(e => !(lookup[e] === undefined || lookup[e] === null))
+
+            if (missingKeys.length > 0) {
+                const localization = await GobchatAPI.getLocalizedStrings(language, missingKeys)
+                if (lookup)
+                    Object.entries(localization).forEach((k, v) => lookup[k] = v)
+                else
+                    lookup = localization
+            }
+
+            var results = lookup
+            if (Object.keys(lookup).length !== keys.length)
+                results = keys.map(e => lookup[e])
+
+            if (params && params.length > 0)
+                for (var key in Object.keys(results))
+                    results[key] = Gobchat.formatString(results[key], params)
+
+            if (keys.length === 1)
+                return results[keys[0]]
+
+            return results
         }
 
         async get(key, language) {